test: isolate throwing iterator class
